Added get_data file
Added get_data file which contains all functions of getting the
